Consolidate console mods and add individual (not unit) tests
