--- lib/__testfixtures__/mods/expressions/remove-console-expressions.input.js
@@ -1,5 +1,0 @@
-const v = 5 || console.log('hello');
-console.log('test');
-if (v && console.log(v)) {
-  console.log('test');
-}
