--- lib/__testfixtures__/mods/expressions/remove-console-expressions.output.js
@@ -1,5 +1,0 @@
-const v = 5 || undefined;
-console.log('test');
-if (v && undefined) {
-  console.log('test');
-}
