--- lib/__testfixtures__/mods/statements/remove-console-statements.input.js
@@ -1,5 +1,0 @@
-console.log('word');
-let v = 5;
-console.log('hello');
-v = 10;
-console.log(v);
