--- lib/__testfixtures__/mods/statements/remove-console-statements.output.js
@@ -1,2 +1,0 @@
-let v = 5;
-v = 10;
