+++ lib/__testfixtures__/remove-console-expressions.input.js
@@ -1,0 +1,5 @@
+const arr = [1, 2, 3];
+arr.forEach(elt => console.log(elt));
+if (null || console.log(arr.length)) {
+  const a = 5 || console.log('test');
+}
