+++ lib/__testfixtures__/remove-console-expressions.output.js
@@ -1,0 +1,5 @@
+const arr = [1, 2, 3];
+arr.forEach(elt => undefined);
+if (null || undefined) {
+  const a = 5 || undefined;
+}
