+++ lib/__testfixtures__/remove-console-statements.input.js
