+++ lib/__testfixtures__/remove-console-statements.output.js
