+++ lib/rewriter.js
@@ -1,12 +1,10 @@
-const cleanConsoleExpressions = require('./mods/clean-console-expressions.js');
-const cleanConsoleStatements = require('./mods/clean-console-statements.js');
+const cleanConsole = require('./mods/clean-console.js');
 
 module.exports = (fileInfo, api, options) => {
   const j = api.jscodeshift;
   const root = j(fileInfo.source);
 
-  cleanConsoleExpressions(j, root);
-  cleanConsoleStatements(j, root);
+  cleanConsole(j, root);
 
   return root.toSource();
 };
