+++ lib/runTestFixtures.js
@@ -8,7 +8,6 @@ const defineTest = require('jscodeshift/src/testUtils').defineTest;
 function runTestFixtures() {
   walkDir(path.join(__dirname, '/__testfixtures__'), (file) => {
     const [filename, ioType] = file.split('.');
-    console.log(filename);
     if (ioType === 'input') {
       defineTest(path.join(__dirname, '__tests__'),
           'rewriter', undefined, filename);
